requirements.txt
Flask==3.1.1
requests==2.32.4
redis==6.4.0
python-dotenv==1.1.1